app-alice/requirements.txt
requests
pyprimes==0.2.2a0
pycrypto
gunicorn
Flask
WTForms
Flask-WTF
jwcrypto